requirements.txt
torch
pytorch-fast-transformers